// frontend/src/types/generated/CentralOwnerSessionStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CentralOwnerSessionEnd } from "./CentralOwnerSessionEnd.js";

export type CentralOwnerSessionStatus = { "state": "active", } | { "state": "ended", reason: CentralOwnerSessionEnd, };